pyproject.toml
# ─────────────────────────────────────────────────────────────────────────────
# Not a distributable package — just central config for tools used in CI
# ─────────────────────────────────────────────────────────────────────────────

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "over-intra-news"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
  "pandas",
  "psycopg2-binary",
  "exchange_calendars",
  "python-dotenv",
  "requests",
  "boto3",
  "lxml",
  "sqlalchemy",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "ruff",
  "black",
  "pandas-stubs",
  "types-psycopg2",
  "types-requests",
  "boto3-stubs",
  "pytest",
  "pytest-mock",
  "lxml-stubs"
]

[tool.mypy]
python_version = "3.13"
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"
files = ["src/infra", "src/aws/ccnews_sampler", "tests"]

[[tool.mypy.overrides]]
module = ["exchange_calendars.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["aws", "infra", "tests"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["notebooks_utils", "infra", "aws"]